+++ tests/unit/plugins/modules/test_cifsserver.py
@@ -14,8 +14,6 @@ from ansible_collections.dellemc.unity.tests.unit.plugins.module_utils.mock_cifs
     import MockCIFSServerApi
 from ansible_collections.dellemc.unity.tests.unit.plugins.module_utils.mock_sdk_response \
     import MockSDKObject
-from ansible_collections.dellemc.unity.tests.unit.plugins.module_utils.mock_api_exception \
-    import HttpError as http_error, MockApiException
 from ansible_collections.dellemc.unity.plugins.module_utils.storage.dell import utils
 
 utils.get_logger = MagicMock()